requirements.txt
azure-cosmos==3.1.2
Flask==1.1.1
geopy